cooldown / 03:44 / 뷰 약간 추가
index.jsp 추가함
resources/assets 추가
구독 비즈니스 로직과 컨트롤러 약간 수정
subscribe.sql 추가

+++ src/main/java/com/teulda/service/community/impl/CommunityServiceImpl.java
@@ -120,7 +120,7 @@ import com.teulda.service.domain.Subscribe;
 
 	@Override
 	public boolean addSubscribe(Subscribe subscribe) throws Exception {
-		if(communityDao.checkSubscribe(subscribe) == 0) {
+		if(communityDao.checkSubscribe(subscribe)) {
 			communityDao.addSubscribe(subscribe);
 			return true;
 		}
@@ -133,10 +133,17 @@ import com.teulda.service.domain.Subscribe;
 	}
 
 	@Override
-	public void deleteSubscribe(Subscribe subscribe) throws Exception {
-		communityDao.deleteSubscribe(subscribe);
-		
+	public boolean deleteSubscribe(Subscribe subscribe) throws Exception {
+		if(!communityDao.checkSubscribe(subscribe)) {
+			communityDao.deleteSubscribe(subscribe);
+			return true;
+		}
+		return false;
 	}
 
+	@Override
+	public boolean checkSubscribe(Subscribe subscribe) throws Exception {
+		return communityDao.checkSubscribe(subscribe);	//	구독 중이면 false, 구독하지 않는 중이면 true.(그러니까 구독 가능하면 true)
+	}
 
 }
